moved functions from tf_utils
